+++ packages/components/pie-assistive-text/test/component/pie-assistive-text.spec.ts
@@ -1,0 +1,20 @@
+
+import { test, expect } from '@sand4rt/experimental-ct-web';
+import { PieAssistiveText, AssistiveTextProps } from '../../src/index.ts';
+
+const componentSelector = '[data-test-id="pie-assistive-text"]';
+
+test.describe('PieAssistiveText - Component tests', () => {
+    test('should render successfully', async ({ mount, page }) => {
+        // Arrange
+        await mount(PieAssistiveText, {
+            props: {} as AssistiveTextProps,
+        });
+
+        // Act
+        const assistiveText = page.locator(componentSelector);
+
+        // Assert
+        expect(assistiveText).toBeVisible();
+    });
+});